Do not add margin light when opening modal on mobile
